+++ CHANGELOG.md
@@ -9,9 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Devices
   - HMIP-FBL (Blind Actuator - flush-mount)
   - HMIP-BRC2 (Remote Control for brand switches – 2 channels)
+  - HmIP-eTRV-C (Heating-thermostat compact without display)
 
 - API
   - AutoNameEnum.from_str does now take a default parameter_ 
+  - HeatingThermostat.**valveActualTemperature** = we are now able to read the measured temperature on the VALVE!
 
 ## [0.10.5] - 2019-01-26
 ### Added